Moved base token filter to internal package
