+++ pmd-core/src/main/java/net/sourceforge/pmd/cpd/token/AntlrTokenFilter.java
@@ -6,6 +6,7 @@ package net.sourceforge.pmd.cpd.token;
 
 import static org.antlr.v4.runtime.Token.EOF;
 
+import net.sourceforge.pmd.cpd.token.internal.BaseTokenFilter;
 import net.sourceforge.pmd.lang.antlr.AntlrTokenManager;
 
 /**
